backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg[binary]>=3.1.0; python_version >= "3.13"
pymongo==4.6.1
python-dotenv==1.0.0
marshmallow==3.20.1
opencv-python==4.8.1.78
numpy>=1.26.0,<2.0.0  # NumPy 2.x incompatible with OpenCV 4.8.1.78
face-recognition==1.3.0
Pillow>=10.1.0
Werkzeug==3.0.1
python-dateutil==2.8.2

